% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_int}
\alias{format_int}
\title{Format integers with thousand separators}
\usage{
format_int(x)
}
\arguments{
\item{x}{Integer-like numeric vector.}
}
\value{
A character vector.
}
\description{
Format integers with thousand separators
}
