% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{numformat}
\alias{numformat}
\title{Format numbers with commas and suppressed leading zero}
\usage{
numformat(value, nsmall = 2)
}
\arguments{
\item{value}{Numeric vector.}

\item{nsmall}{Number of decimals for non-integers.}
}
\value{
A character vector.
}
\description{
Formats numeric values by adding thousand separators, forcing decimals
when needed, and removing leading zeros for values between -1 and 1.
}
