% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{to_percent}
\alias{to_percent}
\title{Convert proportions to percent strings}
\usage{
to_percent(x, digits = 1)
}
\arguments{
\item{x}{Numeric vector of proportions.}

\item{digits}{Number of decimal places.}
}
\value{
A character vector.
}
\description{
Convert proportions to percent strings
}
